refactor: Update image_store table schema to include type column
This commit updates the image_store table schema in the src/server/db/schema.ts file to include a new column called "type". The type column is of type "typeEnum" and has a default value of "clsS". This change improves the database structure by allowing for more specific categorization of image stores. It ensures consistency and flexibility in managing different types of image stores.