feat: Add image-size npm dependency for image manipulation
This commit adds the "image-size" npm dependency to the project. The "image-size" library provides functionality to extract dimensions from various image formats. This dependency will be used to retrieve the width and height of uploaded images and store them in the database schema. It will enhance the application's image manipulation capabilities.