feat: Retrieve and store image dimensions for uploaded images
This commit adds functionality to retrieve the width and height of uploaded images using the "image-size" npm dependency. The retrieved dimensions are then stored in the database schema for the "images" table. This enhancement improves the application's image manipulation capabilities and allows for easier retrieval and manipulation of image data.