feat: Add support for Google Cloud Storage in image upload
This commit adds support for Google Cloud Storage (GCS) in the image upload functionality. It introduces a new dependency on the "@google-cloud/storage" package and modifies the schema of the "images" table in the database to include a new column "gsutilURI" for storing the GCS URI of the uploaded image. Additionally, it introduces new functions "uploadJsonToGCS" and "uploadToGCS" in the "gcs/index.ts" file for uploading JSON and image files to GCS respectively. This enhancement improves the scalability and reliability of the application by leveraging GCS for storing images.