refactor: Update ZoomDialog component to conditionally render Save button
The ZoomDialog component in the src/components/image/zoom-dialog.tsx file is updated to conditionally render the Save button based on the imageStore type. If the imageStore type is either "clsS" or "clsM", the Save button will be displayed. This change improves the user experience by only showing the Save button when it is relevant to the current imageStore type.