refactor: Update columns.tsx to include type and duration fields
The columns.tsx file in the training/_components directory is updated to include the "type" and "duration" fields in the TrainingJob table. The "type" field is displayed as a column header and the "duration" field is calculated and displayed as minutes. This change improves the visibility and analysis of training job information in the table.