refactor: Update imageRouter to include toggleIsLabeled mutation
The imageRouter in the src/server/api/routers/image.ts file is updated to include a new mutation called toggleIsLabeled. This mutation allows the user to toggle the isLabeled property of an image, updating it to the opposite value. This change improves the functionality of the image router by providing a convenient way to change the labeling status of an image.