refactor: Update imageRouter to fix file extension in gcsPath
The imageRouter in image.ts is updated to fix the file extension in the gcsPath variable. Previously, it was set to ".json" instead of ".png". This change ensures that the correct file extension is used when uploading images to Google Cloud Storage (GCS), improving the accuracy and functionality of the code.